# docs/swagger.yaml
basePath: /
definitions:
  handler.AddGroupAnimeRequest:
    properties:
      library_anime_id:
        minimum: 1
        type: integer
    required:
    - library_anime_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API untuk personal anime library.
  title: Libra - Anime Library API
  version: "1.0"
paths:
...
  /me/groups/{id}/animes:
    get:
      description: Get all library anime assigned to a specific group.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Group anime retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid group ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get anime in group
      tags:
      - Group Anime
    post:
      consumes:
      - application/json
      description: Add an anime from the user's library to a group.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Library anime to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.AddGroupAnimeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Anime added to group
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group or library anime not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Anime already exists in group
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add anime to group
      tags:
      - Group Anime
  /me/groups/{id}/animes/{library_anime_id}:
    delete:
      description: Remove an anime from a group without deleting it from the user's
        library.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Library anime ID
        in: path
        name: library_anime_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Anime removed from group
        "400":
          description: Invalid group ID or library anime ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group or anime in group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove anime from group
      tags:
      - Group Anime
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"